packages/running-analysis/pyproject.toml
[project]
name = "running-analysis"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Andy Davies", email = "andrew.m.davies.2@gmail.com" }
]
requires-python = ">=3.11"
dependencies = [
    "pandas-stubs>=3.0.0.260204",
]

[build-system]
requires = ["uv_build>=0.9.29,<0.10.0"]
build-backend = "uv_build"
